Added IMessage.Embeds; Added NotImplementedEmbed.Resolve; Added IPresence; Moved OnlyPressToTalk to OnlyPushToTalk; Moved IUser.IsOnline to IPresence.IsOnline; Fixed SocketUser.Update NRE; Added docs; Bumped into v0.0.20-alpha.

+++ src/KaiHeiLa.Net.Rest/API/Rest/UpdateIntimacyValueParams.cs
@@ -14,5 +14,5 @@ internal class UpdateIntimacyValueParams
 
     [JsonPropertyName("img_id")]
     [JsonNumberHandling(JsonNumberHandling.WriteAsString | JsonNumberHandling.AllowReadingFromString)]
-    public ulong ImageId { get; set; }
+    public uint ImageId { get; set; }
 }
